requirements.txt
Django==4.2.1
tzdata==2023.3
pytz==2022.4
python-dateutil==2.8.2
djangorestframework==3.14.0